Files for Project3a
Added files for project3a. Dockerfile and docker-compose file have been tested and work. Unfortunately, I was not able to get the stocks to generate in the python program using the API so I had to put them in manually by hand in the html file. Other than that, the program does not crash and posts. The chart displays once the user selects the stock information.